+++ src/services/notifications/index.ts
@@ -1,0 +1,3 @@
+export * from './notifications.interface';
+export * from './notificationsOptions';
+export * from './defaultNotifications.service';