+++ src/services/positioning/positioning.interface.ts
@@ -1,0 +1,6 @@
+/**
+ * Allows relative positioning of elements
+ */
+export interface Positioning
+{
+}